+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.19.0.0
 
+* Add `FromJSON` instance to `Vote`
+* Move `Vote` to `BaseTypes` module
 * Add `poolParamsCount`, `decCBORGroupPoolParams`, `encCBORGroupPoolParams` to `StakePool` module
 * Remove `DecCBORGroup` and `EncCBORGroup` instances for `PoolParams`
 * Remove the `UMap` module and the `umap` benchmarks cabal target.